+++ packages/column-views/src/maps/index.ts
@@ -1,3 +1,4 @@
 export * from "./column-navigation";
 export * from "./column-correlation";
 export * from "./_shared";
+export * from "./units-overlay";