+++ packages/data-components/src/typings.d.ts
@@ -1,7 +1,0 @@
-declare module "*.module.sass" {
-  interface IClassNames {
-    [className: string]: string;
-  }
-  const classNames: IClassNames;
-  export = classNames;
-}