+++ packages/map-styles/src/index.ts
@@ -1,5 +1,3 @@
-import "./images.d.ts";
-
 export * from "./layer-helpers";
 export * from "./styles";
 export * from "./style-image-manager";