+++ packages/map-styles/src/styles/index.ts
@@ -1,2 +1,3 @@
 export * from "./line-symbols";
 export * from "./layers";
+export * from "./types";