+++ packages/ui-components/src/index.ts
@@ -23,5 +23,3 @@ export * from "./page-layouts";
 export * from "./dev-panel";
 export * from "./data";
 export * from "./image";
-
-import "./module.d.ts";