+++ toolchain/bundler/src/index.ts
@@ -10,7 +10,7 @@ import {
   readPackageJSON,
 } from "./check-entries.js";
 import chalk from "chalk";
-import { cjsInterop } from "vite-plugin-cjs-interop";
+import hyperStyles from "../../vite-plugin-hyperstyles/src/index.js";
 
 const module = import.meta.url;
 const __file = fileURLToPath(module);
@@ -110,6 +110,7 @@ function buildStandardViteConfig(
         logLevel: verbose ? "info" : "silent",
       }) as any,
       checkExportsPlugin,
+      hyperStyles(),
       // cjsInterop({
       //   dependencies: ["labella", "ui-box"],
       // }),