+++ packages/mapbox-utils/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.2] - 2026-02-06
+
+- Update Mapbox dependency to v3
+- Improved style helper functions
+
 ## [1.7.1] - 2026-01-29
 
 - Change layout of `package.json` and explicitly mark no side effects.